% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_helper.R
\name{can_permute}
\alias{can_permute}
\title{Can Permute Checker}
\usage{
can_permute(extract_func, summary_func, permute_CSR, markvar_levels)
}
\arguments{
\item{extract_func}{univariate or bivariate function}

\item{summary_func}{summary function}

\item{permute_CSR}{whether or not user wants to permute}

\item{markvar_levels}{how many levels are in the column the user wants for column of interest}
}
\value{
\code{TRUE} or \code{FALSE} on if permutations are possible
}
\description{
Can Permute Checker
}
\keyword{internal}
