% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funreg_utils.R
\name{extract_c}
\alias{extract_c}
\title{extract_c}
\usage{
extract_c(fit, survival_time, event)
}
\arguments{
\item{fit}{fit AFCM or LFCM model fit object.}

\item{survival_time}{Vector of survival/censoring times}

\item{event}{Survival statust (0 = censored, 1 = event)}
}
\description{
Function to calculate c-index from a an AFCM or LFCM fit
}
\author{
Erjia Cui
}
