% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_model.R
\name{extract_model}
\alias{extract_model}
\title{Extract Model}
\usage{
extract_model(mxFDAobject, what, model_name)
}
\arguments{
\item{mxFDAobject}{object of class \code{mxFDA}}

\item{what}{what functional PCA data to extract, e.g. 'uni k'}

\item{model_name}{character string of the model name to retrieve}
}
\value{
fit functional model
}
\description{
Currently only extracts functional cox models not mixed functional cox models.
}
\examples{
#set seed
set.seed(333)

}
\author{
Alex Soupir \email{alex.soupir@moffitt.org}
}
