% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_helper.R
\name{filter_data}
\alias{filter_data}
\title{Filter Data}
\usage{
filter_data(dat, filter_columns)
}
\arguments{
\item{dat}{a data frame which to filter}

\item{filter_columns}{named vector where names is the column and element is what to filter to}
}
\value{
data frame filtered
}
\description{
not used in an important want currently
}
\details{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\keyword{internal}
