% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_spatial.R
\name{filter_spatial}
\alias{filter_spatial}
\title{Filter Spatial data}
\usage{
filter_spatial(mxFDAobject, ...)
}
\arguments{
\item{mxFDAobject}{object of class \code{mxFDA}}

\item{...}{expressions that return a logical TRUE/FALSE value when evaluated on columns of the spatial data slot. These expressions get passed to \code{\link[dplyr:filter]{dplyr::filter()}} so must be compatible.}
}
\value{
mxFDAobject with updated spatial slot
}
\description{
function to filter the spatial data slot of the \code{mxFDA} object.
}
\examples{
# example code
set.seed(333)

}
\references{
\code{\link[dplyr:filter]{dplyr::filter()}}
}
\author{
Alex Soupir \email{alex.soupir@moffitt.org}
}
