% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatstat_funcs.R
\name{Gcross}
\alias{Gcross}
\title{Gcross}
\value{
object of class "fv" with columns for radii, theoretical bivariate G, and columns for edge corrections.
}
\description{
See \code{\link[spatstat.explore:Gcross]{spatstat.explore::Gcross()}}
}
\keyword{internal}
