% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatstat_funcs.R
\name{Gest}
\alias{Gest}
\title{Gest}
\value{
object of class "fv" with columns for radii, theoretical G, and columns for edge corrections.
}
\description{
See \code{\link[spatstat.explore:Gest]{spatstat.explore::Gest()}}
}
\keyword{internal}
