% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_helper.R
\name{get_data}
\alias{get_data}
\title{Get Data}
\usage{
get_data(mxFDAobject, what, type)
}
\arguments{
\item{mxFDAobject}{object of class \code{mxFDA}}

\item{what}{character string of length 2 for what to return. produced by unlisting "uni k" or "bi g fpca" type strings}

\item{type}{the type of data to get for the what summary function}
}
\value{
slot data
}
\description{
Returns data from slots within the mxfDA object
}
\details{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}
}
\keyword{internal}
