% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_helper.R
\name{is.empty}
\alias{is.empty}
\title{Is Empty}
\usage{
is.empty(obj, slot.name)
}
\arguments{
\item{obj}{object with slots}

\item{slot.name}{character string of for the name of a slot}
}
\value{
\code{TRUE}, \code{FALSE}, or \code{stop} command
}
\description{
checks whether or not a slot is empty in the mxfDA objecct
}
\details{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}
}
\keyword{internal}
