% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatstat_funcs.R
\name{Kcross}
\alias{Kcross}
\title{Kcross}
\value{
object of class "fv" with columns for radii, theoretical bivariate K, and columns for edge corrections.
}
\description{
See \code{\link[spatstat.explore:Kcross]{spatstat.explore::Kcross()}}
}
\keyword{internal}
