% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatstat_funcs.R
\name{Kest}
\alias{Kest}
\title{Kest}
\value{
object of class "fv" with columns for radii, theoretical K, and columns for edge corrections.
}
\description{
See \code{\link[spatstat.explore:Kest]{spatstat.explore::Kest()}}
}
\keyword{internal}
