% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatstat_funcs.R
\name{Lcross}
\alias{Lcross}
\title{Lcross}
\value{
object of class "fv" with columns for radii, theoretical bivariate L, and columns for edge corrections.
}
\description{
See \code{\link[spatstat.explore:Lcross]{spatstat.explore::Lcross()}}
}
\keyword{internal}
