% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatstat_funcs.R
\name{Lest}
\alias{Lest}
\title{Lest}
\value{
object of class "fv" with columns for radii, theoretical L, and columns for edge corrections.
}
\description{
See \code{\link[spatstat.explore:Lest]{spatstat.explore::Lest()}}
}
\keyword{internal}
