% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lung_df}
\alias{lung_df}
\title{Multiplex imaging data from a non-small cell lung cancer study.}
\format{
\subsection{\code{lung_df}}{

A data frame with 879,694 rows and 19 columns:
\describe{
\item{image_id}{Image id for a given patient}
\item{patient_id}{Unique patient id}
\item{age}{Patient age at time of cancer diagnosis}
\item{survival_days}{Survival time from diagnosis, in days}
\item{survival_status}{Censoring variable, 1 = death, 0 = censor}
\item{x}{Cell x position}
\item{y}{Cell y position}
...
}
}
}
\source{
\url{https://bioconductor.org/packages/release/data/experiment/html/VectraPolarisData.html}
}
\usage{
lung_df
}
\description{
This data is adapted from the VectraPolarisData Bioconductor package. There are multiple ROIs for each patient.
}
\keyword{datasets}
