% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_helper.R
\name{metric.exists}
\alias{metric.exists}
\title{Check if Spatial Summary Exists}
\usage{
metric.exists(mxFDAobject, metric)
}
\arguments{
\item{mxFDAobject}{object of class \code{mxFDA}}

\item{metric}{a character vector split with and passed by other functions}
}
\value{
a stopped function if slot not in summary functions
}
\description{
Makes sure in processing functions that the spatial summary function has been calculated before running FPCA or models
}
\details{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}
}
\keyword{internal}
