% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxfda-package.R
\docType{package}
\name{mxfda-package}
\alias{mxfda}
\alias{mxfda-package}
\title{mxfda: A Functional Data Analysis Package for Spatial Single Cell Data}
\description{
Methods and tools for deriving spatial summary functions from single-cell imaging data and performing functional data analyses. Functions can be applied to other single-cell technologies such as spatial transcriptomics. Functional regression and functional principal component analysis methods are in the 'refund' package \url{https://cran.r-project.org/package=refund} while calculation of the spatial summary functions are from the 'spatstat' package \url{https://spatstat.org/}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/julia-wrobel/mxfda/}
  \item \url{http://juliawrobel.com/mxfda/}
  \item Report bugs at \url{https://github.com/julia-wrobel/mxfda/issues/}
}

}
\author{
\strong{Maintainer}: Alex Soupir \email{alex.soupir@moffitt.org} (\href{https://orcid.org/0000-0003-1251-9179}{ORCID})

Authors:
\itemize{
  \item Julia Wrobel \email{julia.wrobel@emory.edu} (\href{https://orcid.org/0000-0001-6783-1421}{ORCID})
}

}
\keyword{internal}
