% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_helper.R
\name{one_zero}
\alias{one_zero}
\title{1/0 Checker}
\usage{
one_zero(vec)
}
\arguments{
\item{vec}{vector of hopefully 1s and 0s but anything}
}
\value{
\code{TRUE} or \code{FALSE} depending on if something other than 0/1 is in the vector
}
\description{
1/0 Checker
}
\details{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}
}
\keyword{internal}
