% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ovarian_df}
\alias{ovarian_df}
\title{Multiplex imaging data from an ovarian cancer tumor microarray}
\format{
\subsection{\code{ovarian_df}}{

A data frame with 108,608 rows and 7 columns:
\describe{
\item{patient_id}{Unique patient id}
\item{age}{Patient age at time of cancer diagnosis}
\item{survival_time}{Survival time from diagnosis, in days}
\item{event}{Censoring variable, 1 = death, 0 = censor}
\item{x}{Cell x position}
\item{y}{Cell y position}
\item{immune}
}
}
}
\source{
\url{https://bioconductor.org/packages/release/data/experiment/html/VectraPolarisData.html}
}
\usage{
ovarian_df
}
\description{
This data is adapted from the VectraPolarisData Bioconductor package.
Signal between the outcomes and spatial summary functions has been augmented for teaching purposes.
}
\keyword{datasets}
