% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ovarian_FDA}
\alias{ovarian_FDA}
\title{Multiplex imaging data from an ovarian cancer tumor microarray}
\format{
\subsection{\code{ovarian_FDA}}{

A data frame with 12,726 rows and 8 columns:
\describe{
\item{patient_id}{Unique patient id}
\item{age}{Patient age at time of cancer diagnosis}
\item{survival_time}{Survival time from diagnosis, in days}
\item{event}{Censoring variable, 1 = death, 0 = censor}
\item{r}{radius}
\item{sumfun}{Value of G function}
\item{csr}{Value of G function under assumption of complete spatial randomness}
\item{fundiff}{sumfun-csr}
}
}
}
\source{
\url{https://bioconductor.org/packages/release/data/experiment/html/VectraPolarisData.html}
}
\usage{
ovarian_FDA
}
\description{
This data is adapted from the VectraPolarisData Bioconductor package and comes from a tumor-microarray of tissue samples from 128 patients with ovarian cancer.
There is one patient per subject.
}
\details{
Spatial summary functions of ovarian cancer multiplex imaging data.

This data is adapted from the VectraPolarisData Bioconductor package.
Signal between the survival outcome and spatial summary functions has been augmented for teaching purposes.
Spatial relationship is summarized using the nearest neighbor G function.
}
\keyword{datasets}
