% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funreg_utils.R
\name{pfr2}
\alias{pfr2}
\title{pfr2}
\usage{
pfr2(formula = NULL, fitter = NA, method = "REML", ...)
}
\description{
hot-fix for the prf function from refund::
}
\details{
Assumes users wont use gamm4 or lme4 packages for dependency purposes

see \code{?refund::pfv}
}
\author{
Jonathan Geller

Edited by Alex Soupir \email{alex.soupir@moffitt.org}
}
\keyword{internal}
