% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.afcmSurface}
\alias{plot.afcmSurface}
\title{Plot afcm object}
\usage{
\method{plot}{afcmSurface}(x, ...)
}
\arguments{
\item{x}{object of class \code{afcmSurface} to be plotted}

\item{...}{currently ignored}
}
\value{
object compatable with ggplot2
}
\description{
Plot afcm object
}
\examples{
#set seed
set.seed(333)

}
\author{
Julia Wrobel \email{julia.wrobel@emory.edu}

Alex Soupir \email{alex.soupir@moffitt.org}
}
