% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.lfcmSurface}
\alias{plot.lfcmSurface}
\title{Plot lfcm surface}
\usage{
\method{plot}{lfcmSurface}(x, ...)
}
\arguments{
\item{x}{object of class \code{lfcmSurface} to be plotted}

\item{...}{currently ignored}
}
\value{
object compatable with ggplot2
}
\description{
Plot lfcm surface
}
\author{
Julia Wrobel \email{julia.wrobel@emory.edu}

Alex Soupir \email{alex.soupir@moffitt.org}
}
