% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.sofr}
\alias{plot.sofr}
\title{Plot sofr object}
\usage{
\method{plot}{sofr}(x, ...)
}
\arguments{
\item{x}{object of class \code{sofr} to be plotted}

\item{...}{currently ignored}
}
\value{
object compatable with ggplot2
}
\description{
Plot sofr object
}
\author{
Julia Wrobel \email{julia.wrobel@emory.edu}

Alex Soupir \email{alex.soupir@moffitt.org}
}
