% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_fpc.R
\name{plot_fpc}
\alias{plot_fpc}
\title{Create plot of mean +/- scaled eigenfunctions from FPCA}
\usage{
plot_fpc(obj, pc_choice)
}
\arguments{
\item{obj}{fpca object to be plotted.}

\item{pc_choice}{FPC to be plotted.}
}
\value{
object of class ggplot
}
\description{
Produces a ggplot with mean plus or minus two standard deviations of a selected FPC.
}
\author{
Julia Wrobel \email{julia.wrobel@emory.edu}
}
