% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatstat_funcs.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{Kest}
\alias{Gest}
\alias{Lest}
\alias{Kcross}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{spatstat.explore}{\code{\link[spatstat.explore]{Gest}}, \code{\link[spatstat.explore]{Kcross}}, \code{\link[spatstat.explore]{Kcross}}, \code{\link[spatstat.explore]{Kcross}}, \code{\link[spatstat.explore]{Kest}}, \code{\link[spatstat.explore]{Lest}}}
}}

