% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_fpca.R
\name{run_fpca}
\alias{run_fpca}
\title{run_fpca}
\usage{
run_fpca(
  mxfundata,
  id,
  r = "r",
  value = "fundiff",
  knots = NULL,
  analysis_vars = NULL,
  lightweight = FALSE,
  ...
)
}
\arguments{
\item{mxfundata}{Dataframe of spatial summary functions from multiplex imaging data, in long format. Can be estimated using the function \code{extract_summary_functions} or provided separately. For FPCA there should be one function per image, and only one image per subject.}

\item{id}{Character string, the name of the variable that identifies each unique subject.}

\item{r}{Character string, the name of the variable that identifies the function domain (usually a radius for spatial summary functions). Default is "r".}

\item{value}{Character string, the name of the variable that identifies the spatial summary function values. Default is "fundiff".}

\item{knots}{Number of knots for defining spline basis.Defaults to the number of measurements per function divided by 2.}

\item{analysis_vars}{Optional list of variables to be retained for downstream analysis.}

\item{lightweight}{Default is FALSE. If TRUE, removes Y and Yhat from returned FPCA object. A good option to select for large datasets.}

\item{...}{Optional other arguments to be passed to \code{fpca.face}}
}
\value{
A \code{list} containing:
\item{mxfundata}{The original dataframe of spatial summary functions, with scores from FPCA appended for downstream modeling}
\item{fpc_object}{A list of class "fpca" with elements described in the documentation for  \code{refund::fpca.face}}
}
\description{
This is a wrapper for the function \code{fpca.face} from the \code{refund} package. EXPAND
}
\examples{
# simulate data
set.seed(1001)


}
\references{
Xiao, L., Ruppert, D., Zipunnikov, V., and Crainiceanu, C. (2016).
Fast covariance estimation for high-dimensional functional data.
\emph{Statistics and Computing}, 26, 409-421.
DOI: 10.1007/s11222-014-9485-x.
}
