% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.mxFDA}
\alias{summary.mxFDA}
\title{Summary method for object of class \code{mxFDA}}
\usage{
\method{summary}{mxFDA}(x, ...)
}
\arguments{
\item{x}{object of class \code{mxFDA}}

\item{...}{unused currently}
}
\value{
summary of object
}
\description{
Summary method for object of class \code{mxFDA}
}
\examples{
#set seed
set.seed(333)

}
\author{
Alex Soupir \email{alex.soupir@moffitt.org}
}
