% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_entropy.R
\name{entropy}
\alias{entropy}
\title{Entropy}
\usage{
entropy(X, i, j, r_vec, correction)
}
\arguments{
\item{X}{object of class \code{ppp} from spatstat with 2 marks}

\item{i}{ignored}

\item{j}{ignored}

\item{r_vec}{vector of length wanted for breaks (will be rescaled) with max value at max for measuring entropy}

\item{correction}{ignored}
}
\value{
data frame with entropy calculated for \code{length(r_vec)} bins within 0 to \code{max(r_vec)}
}
\description{
Entropy
}
\details{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\references{
Vu, T., Seal, S., Ghosh, T., Ahmadian, M., Wrobel, J., & Ghosh, D. (2023).
FunSpace: A functional and spatial analytic approach to cell imaging data using entropy measures.
\emph{PLOS Computational Biology}, 19(9), e1011490.

Altieri, L., Cocchi, D., & Roli, G. (2018).
A new approach to spatial entropy measures.
\emph{Environmental and ecological statistics}, 25, 95-110.
}
\author{
Thao Vu \email{thao.3.vu@cuanschutz.edu}

Alex Soupir \email{alex.soupir@moffitt.org}
}
