% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_fpca_object.R
\name{extract_fpca_object}
\alias{extract_fpca_object}
\title{Extract FPCA object}
\usage{
extract_fpca_object(mxFDAobject, what)
}
\arguments{
\item{mxFDAobject}{object of class \code{mxFDA}}

\item{what}{what functional PCA data to extract, e.g. 'uni k'}
}
\value{
fpca object
}
\description{
Extract FPCA object
}
\examples{
#set seed
set.seed(333)

}
\author{
Alex Soupir \email{alex.soupir@moffitt.org}
}
