% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_fpca_object.R
\name{extract_fpca_object}
\alias{extract_fpca_object}
\title{Extract FPCA object}
\usage{
extract_fpca_object(mxFDAobject, what, id, r, value)
}
\arguments{
\item{mxFDAobject}{object of class \code{mxFDA}}

\item{what}{what functional PCA data to extract, e.g. 'uni k'}

\item{id}{sample id column used in \code{run_fpca} function}

\item{r}{column id for radii}

\item{value}{value used to calculate fpcs}
}
\value{
fpca object
}
\description{
Extract FPCA object
}
