% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_spatial_summary.R
\name{extract_spatial_summary}
\alias{extract_spatial_summary}
\title{Summarise spatial data in mxFDA object}
\usage{
extract_spatial_summary(mxFDAobject, columns, grouping_columns = NULL)
}
\arguments{
\item{mxFDAobject}{object of class \code{mxFDA}}

\item{columns}{character vector for column heading for cells to summarise}

\item{grouping_columns}{character vector of other columns to use as grouping, such as region classification column}
}
\value{
data frame with percent of total points per spatial sample \code{column} are positive for
}
\description{
Summarise spatial data in mxFDA object
}
\examples{
#set seed
set.seed(333)

}
\author{
Alex Soupir \email{alex.soupir@moffitt.org}
}
