% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_summary_functions.R
\name{extract_summary_functions}
\alias{extract_summary_functions}
\title{extract_summary_functions}
\usage{
extract_summary_functions(
  mxFDAobject,
  r_vec = seq(0, 100, by = 10),
  extract_func = c(extract_univariate, extract_bivariate),
  summary_func = c(Kest, Lest, Gest),
  markvar,
  mark1,
  mark2 = NULL,
  edge_correction
)
}
\arguments{
\item{r_vec}{Numeric vector of radii over which to evaluate spatial summary functions. Must begin at 0.}

\item{extract_func}{Defaults to extract_univariate, which calculates univariate spatial summary functions. Choose extract_bivariate for bivariate spatial summary functions.}

\item{summary_func}{Spatial summary function to calculate. Options are c(Kest, Lest, Gest) which denote Ripley's K, Besag's L, and nearest neighbor G function, respectively.}

\item{markvar}{The name of the variable that denotes cell type(s) of interest. Character.}

\item{mark1}{Character string that denotes first cell type of interest.}

\item{mark2}{Character string that denotes second cell type of interest for calculating bivariate summary statistics. Not used when calculating univariate statistics.}

\item{edge_correction}{Character string that denotes the edge correction method for spatial summary function. For Kest and Lest choose one of c("border", "isotropic", "Ripley", "translate", "none"). For Gest choose one of c("rs", "km", "han")}

\item{mxdata}{Dataframe of cell-level multiplex imaging data. Should have variables x and y to denote x and y spatial locations of each cell.}

\item{analysis_vars}{Optional list of variables to be retained for downstream analysis.}
}
\value{
A \code{data.frame} containing:
\item{image_patient_id}{the unique image id}
\item{r}{the radius of values over which the spatial summary function is evaluated}
\item{sumfun}{the values of the spatial summary function}
\item{csr}{the values of the spatial summary function under complete spatial randomness}
\item{fundiff}{sumfun - csr, positive values indicate clustering and negative values repulsion}
}
\description{
This is the main function for the fastGFPCA package.
The function requires input data \code{Y} to be a dataframe in long format with variables
\code{id}, \code{index}, and \code{value} to indicate subject IDs,
observation times on the domain, and observations, respectively.
The \code{index} must contain the same, equally spaced grid points for each subject.
The number of functional principal components (FPCs) can either be specified
directly (argument \code{npc}) or chosen based on the explained share of
variance (\code{pve}). Families and link functions can be specified using the same
syntax as the \code{family} argument from the \code{lme4::glmer()} function.
}
\examples{
# simulate data
set.seed(1001)


}
\references{
Xiao, L., Ruppert, D., Zipunnikov, V., and Crainiceanu, C. (2016).
Fast covariance estimation for high-dimensional functional data.
\emph{Statistics and Computing}, 26, 409-421.
DOI: 10.1007/s11222-014-9485-x.
}
\author{
Julia Wrobel \email{julia.wrobel@cuanschutz.edu}
}
