% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_surface.R
\name{extract_surface}
\alias{extract_surface}
\title{extract_surface}
\usage{
extract_surface(
  mxfundata,
  fit,
  r = "r",
  value = "fundiff",
  grid_length = 100,
  analysis_vars,
  se = FALSE
)
}
\arguments{
\item{mxfundata}{Dataframe of spatial summary functions from multiplex imaging data, in long format.}

\item{fit}{AFCM or LFCM model fit object}

\item{r}{Character string, the name of the variable that identifies the function domain (usually a radius for spatial summary functions). Default is "r".}

\item{value}{Character string, the name of the variable that identifies the spatial summary function values. Default is "fundiff".}

\item{grid_length}{Length of grid on which to evaluate coefficient functions.}

\item{analysis_vars}{Other variables used in modeling FCM fit.}

\item{se}{Defaults to FALSE, and returns surface. If TRUE, returns coefficient function with standard errors for LFCM, or coefficient surface for AFCM with non NA values that are statistically significant}
}
\value{
A \code{dataframe} with predicted surface for AFCM and LFCM fits for use in plotting
}
\description{
Internal function called by \code{TITLE: regression function} that transforms long format functional data for use in a linear or additive functional Cox model.
}
\examples{
# simulate data
set.seed(1001)


}
