% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_surface.R
\name{extract_surface}
\alias{extract_surface}
\title{extract_surface}
\usage{
extract_surface(
  mxFDAobject,
  metric,
  model = NULL,
  r = "r",
  value = "fundiff",
  grid_length = 100,
  analysis_vars,
  p = 0.05,
  filter_cols = NULL
)
}
\arguments{
\item{mxFDAobject}{object of class \code{mxFDA} with model \code{model} calculated wihtin}

\item{metric}{spatial summary function to extract surface for}

\item{model}{character string for the name of the model for \code{metric} data}

\item{r}{Character string, the name of the variable that identifies the function domain (usually a radius for spatial summary functions). Default is "r".}

\item{value}{Character string, the name of the variable that identifies the spatial summary function values. Default is "fundiff".}

\item{grid_length}{Length of grid on which to evaluate coefficient functions.}

\item{analysis_vars}{Other variables used in modeling FCM fit.}

\item{p}{numeric p-value used for predicting significant AFCM surface}

\item{filter_cols}{a named vector of factors to filter summary functions to in \code{c(Derived_Column = "Level_to_Filter")} format}
}
\value{
A \code{dataframe} with predicted surface for AFCM and LFCM fits for use in plotting
}
\description{
Internal function called by \code{TITLE: regression function} that transforms long format functional data for use in a linear or additive functional Cox model.
}
\examples{
# simulate data
set.seed(1001)

}
\author{
Julia Wrobel \email{julia.wrobel@emory.edu}

Alex Soupir \email{alex.soupir@moffitt.org}
}
