% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxfda_object.R
\name{make_mxfda}
\alias{make_mxfda}
\title{Make mxFDA class object}
\usage{
make_mxfda(metadata, spatial, subject_key, sample_key)
}
\arguments{
\item{metadata}{metadata with column \code{key} for samples}

\item{spatial}{spatial information, either list or df, with column \code{key}}

\item{subject_key}{column name for subject ID}

\item{sample_key}{column linking metadata to spatial data}
}
\value{
object of class mxFDA
}
\description{
Make mxFDA class object
}
