% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_mxfda.R
\name{make_mxfda}
\alias{make_mxfda}
\title{Make mxFDA class object}
\usage{
make_mxfda(metadata, spatial = NULL, subject_key, sample_key)
}
\arguments{
\item{metadata}{metadata with column \code{key} for samples}

\item{spatial}{spatial information, either list or df, with column \code{key}. \code{spatial} can be empty if inputting data already derived}

\item{subject_key}{column name for subject ID}

\item{sample_key}{column linking metadata to spatial data}
}
\value{
object of class mxFDA
}
\description{
Used to create an object of class \code{mxFDA} that can be used with the \link{mxfda} package for functional data analysis.
}
\details{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}
}
\examples{
#select sample metadata
clinical = lung_df \%>\%
  dplyr::select(image_id, patient_id, patientImage_id,
                gender, age, survival_days, survival_status, stage) \%>\%
  dplyr::distinct()
#select the spatial information
spatial = lung_df \%>\%
  dplyr::select(-image_id, -gender, -age, -survival_days, -survival_status, -stage)
sample_id_column = "patientImage_id"
#create the mxFDA object
mxFDAobject = make_mxfda(metadata = clinical,
                         spatial = spatial,
                         subject_key = "patient_id",
                         sample_key = sample_id_column)

}
\author{
Alex Soupir \email{alex.soupir@moffitt.org}
}
