% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxfda-package.R
\docType{package}
\name{mxfda-package}
\alias{mxfda}
\alias{mxfda-package}
\title{mxfda: What the Package Does (Title Case- stuff)}
\description{
More about what it does (maybe more than one line) Use four spaces when indenting paragraphs within the Description.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/julia-wrobel/mxfda}
  \item \url{https://julia-wrobel.github.io/mxfda}
  \item Report bugs at \url{https://github.com/julia-wrobel/mxfda/issues}
}

}
\author{
\strong{Maintainer}: Julia Wrobel \email{julia.wrobel@cuanschutz.edu} (\href{https://orcid.org/0000-0001-6783-1421}{ORCID})

Authors:
\itemize{
  \item Alex Soupir (\href{https://orcid.org/0000-0003-1251-9179}{ORCID})
}

}
\keyword{internal}
