% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mxfda-package.R
\docType{package}
\name{mxfda-package}
\alias{mxfda}
\alias{mxfda-package}
\title{mxfda: A functional data analysis package for spatial single cell data}
\description{
The `mxfda` package contains tools for analyzing spatial single cell data using methods from functional data analysis. While examples in vignettes illustrate the use of the package for single-cell multiplex imaging data, the package can handle analysis of other types of point process data as well, including repeated samples of spatial transcriptomics and geospatial data. Analyses for this package are executed and stored using an S4 object of class `mxFDA`.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/julia-wrobel/mxfda}
  \item \url{http://juliawrobel.com/mxfda}
  \item Report bugs at \url{https://github.com/julia-wrobel/mxfda/issues}
}

}
\author{
\strong{Maintainer}: Alex Soupir (\href{https://orcid.org/0000-0003-1251-9179}{ORCID})

Authors:
\itemize{
  \item Julia Wrobel \email{julia.wrobel@emory.edu} (\href{https://orcid.org/0000-0001-6783-1421}{ORCID})
}

}
\keyword{internal}
