% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.mxFDA}
\alias{plot.mxFDA}
\title{Plot mxFDA object}
\usage{
\method{plot}{mxFDA}(x, ...)
}
\arguments{
\item{x}{object of class \code{mxFDA} to be plotted}

\item{...}{additional paramters including \code{y}, \code{what}, and \code{sampleID} to inform whats to be plotted}
}
\value{
object compatable with ggplot2
}
\description{
Plot mxFDA object
}
