% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_mfpc.R
\name{plot_mfpc}
\alias{plot_mfpc}
\title{Create plot of mean +/- scaled eigenfunctions from FPCA}
\usage{
plot_mfpc(obj, pc_choice_level1, pc_choice_level2)
}
\arguments{
\item{obj}{fpca object to be plotted.}

\item{pc_choice_level1, pc_choice_level2}{FPC to be plotted.}
}
\value{
list of objects of class ggplot
}
\description{
Produces a ggplot with mean plus or minus two standard deviations of a selected FPC.
}
\author{
Julia Wrobel \email{julia.wrobel@emory.edu}
}
