% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.mxFDA}
\alias{summary.mxFDA}
\title{Summary method for object of class \code{mxFDA}}
\usage{
\method{summary}{mxFDA}(object, ...)
}
\arguments{
\item{object}{object of class \code{mxFDA}}

\item{...}{unused currently}
}
\value{
summary of object to the R console
}
\description{
Summary method for object of class \code{mxFDA}
}
\details{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}
}
\author{
Alex Soupir \email{alex.soupir@moffitt.org}
}
